+++ PufferyKit/Sources/RegistrationModule/EmailTextField.swift
@@ -3,11 +3,12 @@ import DesignSystem
 import ComposableArchitecture
 
 struct EmailTextField: View {
-    var onFinish: () -> Void
+    @Binding var email: String
+    let onFinish: () -> Void
     let viewModel: ComposableArchitecture.ViewStore<RegistrationState, RegistrationAction>
 
     var body: some View {
-        TextField("GettingStarted.Login.EmailPlaceholder", text: Binding(get: { viewModel.email }, set: { viewModel.send(.updateEmail($0)) }), onCommit: { viewModel.send(.shouldLogin(onFinish: self.onFinish)) })
+        TextField("GettingStarted.Login.EmailPlaceholder", text: $email, onCommit: { viewModel.send(.shouldLogin(onFinish: self.onFinish)) })
             .textFieldStyle(RoundedTextFieldStyle())
             .keyboardType(.emailAddress)
             .textContentType(.emailAddress)